refactor: 카드 덱의 테스트, 확장성을 위한 리팩토링
