--- src/main/kotlin/blackjack/model/RandomShuffleCardDeck.kt
@@ -1,7 +1,0 @@
-package blackjack.model
-
-class RandomShuffleCardDeck : ShuffleCardDeck {
-    override fun <T> shuffle(deck: MutableList<T>): MutableList<T> {
-        return deck.shuffled().toMutableList()
-    }
-}
